% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cultureResult.R
\name{createCultureResultTable}
\alias{createCultureResultTable}
\title{Run drugPathway package}
\usage{
createCultureResultTable(
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortId,
  minCellCount = 5
)
}
\arguments{
\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{minCellCount}{}
}
\description{
Run drugPathway package
}
\details{
Run the drugPathway package, which implements the cohort pathway design.
}
