% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPathway.R
\name{drugTable1}
\alias{drugTable1}
\title{get table1 about drug exposure}
\usage{
drugTable1(
  drugExposureData = drugExposureData,
  cohortDatabaseSchema = cohortDatabaseSchema,
  cohortTable = cohortTable,
  cohortId = cohortId
)
}
\arguments{
\item{drugExposureData}{Data extracted from drug exposure table}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{cohortId}{cohort Ids}

\item{conceptSets}{concept set json files}
}
\description{
get table1 about drug exposure
}
\details{
Extract concept_id from the json file of concept set
}
