% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the cohort pathway study}
\usage{
execute(
  connectionDetails,
  databaseId,
  databaseName,
  databaseDescription,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  oracleTempSchema,
  cohortTable,
  outputFolder,
  createCohorts = T,
  runPathway = T,
  packageResults = T
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{databaseName}{A string representing a shareable name of your databasd}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes}

\item{createCohorts}{Whether to create the cohorts for the study}

\item{runPathway}{Whether to run the treatment pathway visualization}

\item{packageResults}{Whether to package the results (after removing sensitive details)}

\item{cohortDiagnostics}{Whether to run cohortDiagnostics package}
}
\description{
Execute the cohort pathway study
}
\details{
This function will execute the sepcified parts of the study
}
